Support Numpy 2 via supporting Cirq 1.5+ (#1761)
## Changes for numpy
 - data_types.py -- np.uint8 is stricter now
 - and_bloq.by -- there were np.uint8's sneaking in
 - There are new mypy complaints that I've generally ignored


## Changes for Cirq

We require a more recent version of Cirq to support Numpy 2. 

### ControlledGate and `_has_unitary_`

I've monkeypatched `cirq.ControlledGate`. It was using
`cirq.has_unitary` to determine whether something could be controlled.
The rest of Cirq now treats "has unitary" as the ability to create a
unitary matrix rather than saying whether the gate is reversible.
 - Tracked in #1764 .

### Simulation

There is something fundamentally incompatible between Qualtran's cirq
interop functionality and recent Cirq versions. I've taken the "nuclear
option", and now when you request the cirq simulation of a
`BloqAsCirqGate` it will be done entirely in Qualtran and export a large
unitary matrix to the cirq simulator; which can cause performance
regressions;
 - see #1763.

In this PR, I've implemented the quick fixes to speed up Qualtran
simulation of narrow, deep circuits from my comment on #1336

Some of the unit tests now cause SIGKILL on my machine. I don't know
what's going on here. I've xfailed them.
 - See #1763.


## Related issues

 - #237 
- #1488 (technically fixes, as written; but see performance regression
issue)
 - #1495
 - #1336

+++ qualtran/_infra/data_types.py
@@ -331,7 +331,7 @@ class QIntOnesComp(QDType):
 
     def from_bits(self, bits: Sequence[int]) -> int:
         x = QUInt(self.bitsize).from_bits([b ^ bits[0] for b in bits[1:]])
-        return (-1) ** bits[0] * x
+        return (-1) ** int(bits[0]) * x
 
     def get_classical_domain(self) -> Iterable[int]:
         max_val = 1 << (self.bitsize - 1)
@@ -398,7 +398,7 @@ class QUInt(QDType):
     def from_bits(self, bits: Sequence[int]) -> int:
         return int("".join(str(x) for x in bits), 2)
 
-    def from_bits_array(self, bits_array: NDArray[np.uint8]) -> NDArray[np.integer]:
+    def from_bits_array(self, bits_array: NDArray[np.uint8]) -> NDArray[np.uint64]:
         bitstrings = np.atleast_2d(bits_array)
         if bitstrings.shape[1] != self.bitsize:
             raise ValueError(f"Input bitsize {bitstrings.shape[1]} does not match {self.bitsize=}")
@@ -408,7 +408,7 @@ class QUInt(QDType):
             return super().from_bits_array(bits_array)
 
         basis = 2 ** np.arange(self.bitsize - 1, 0 - 1, -1, dtype=np.uint64)
-        return np.sum(basis * bitstrings, axis=1, dtype=np.uint64)
+        return np.sum(basis * bitstrings, axis=1, dtype=np.uint64)  # type: ignore[return-value]
 
     def assert_valid_classical_val(self, val: int, debug_str: str = 'val'):
         if not isinstance(val, (int, np.integer)):
